requirements.txt
django_otp
rest_framework
djoser
djangorestframework-jwt
channels_redis